+++ src/usr/help.rs
@@ -18,7 +18,7 @@ fn help_command(cmd: &str) -> usr::shell::ExitCode {
 }
 
 fn help_unknown(cmd: &str) -> usr::shell::ExitCode {
-    println!("Help not found for command '{}'", cmd);
+    eprintln!("Help not found for command '{}'", cmd);
     usr::shell::ExitCode::CommandError
 }