+++ src/usr/tcp.rs
@@ -25,7 +25,7 @@ pub fn main(args: &[&str]) -> usr::shell::ExitCode {
     }
 
     if args.len() != 3 {
-        println!("Usage: tcp <host> <port>");
+        eprintln!("Usage: tcp <host> <port>");
         return usr::shell::ExitCode::CommandError;
     }
 
@@ -41,7 +41,7 @@ pub fn main(args: &[&str]) -> usr::shell::ExitCode {
                 ip_addr
             }
             Err(e) => {
-                println!("Could not resolve host: {:?}", e);
+                eprintln!("Could not resolve host: {:?}", e);
                 return usr::shell::ExitCode::CommandError;
             }
         }
@@ -60,11 +60,11 @@ pub fn main(args: &[&str]) -> usr::shell::ExitCode {
     if let Some(ref mut iface) = *sys::net::IFACE.lock() {
         match iface.ipv4_addr() {
             None => {
-                println!("Interface not ready");
+                eprintln!("Interface not ready");
                 return usr::shell::ExitCode::CommandError;
             }
             Some(ip_addr) if ip_addr.is_unspecified() => {
-                println!("Interface not ready");
+                eprintln!("Interface not ready");
                 return usr::shell::ExitCode::CommandError;
             }
             _ => {}
@@ -74,11 +74,11 @@ pub fn main(args: &[&str]) -> usr::shell::ExitCode {
         let started = syscall::realtime();
         loop {
             if syscall::realtime() - started > timeout {
-                println!("Timeout reached");
+                eprintln!("Timeout reached");
                 return usr::shell::ExitCode::CommandError;
             }
             if sys::console::end_of_text() {
-                println!();
+                eprintln!();
                 return usr::shell::ExitCode::CommandError;
             }
             let timestamp = Instant::from_millis((syscall::realtime() * 1000.0) as i64);
@@ -98,7 +98,7 @@ pub fn main(args: &[&str]) -> usr::shell::ExitCode {
                         let local_port = 49152 + random::get_u16() % 16384;
                         println!("Connecting to {}:{}", address, port);
                         if socket.connect((address, port), local_port).is_err() {
-                            println!("Could not connect to {}:{}", address, port);
+                            eprintln!("Could not connect to {}:{}", address, port);
                             return usr::shell::ExitCode::CommandError;
                         }
                         State::Request